+++ src/main/kotlin/it/czerwinski/intellij/wavefront/settings/ObjPreviewSettingsState.kt
@@ -36,6 +36,7 @@ data class ObjPreviewSettingsState(
     @field:Attribute var lineWidth: Float = DEFAULT_LINE_WIDTH,
     @field:Attribute var pointSize: Float = DEFAULT_POINT_SIZE,
     @field:Attribute var cropTextures: Boolean = DEFAULT_CROP_TEXTURES,
+    @field:Attribute var mipmapping: Boolean = DEFAULT_MIPMAPPING,
     @field:Attribute var shaderQuality: ShaderQuality = ShaderQuality.DEFAULT,
     @field:Attribute var displacementQuality: Float = DEFAULT_DISPLACEMENT_QUALITY,
 ) {
@@ -53,6 +54,7 @@ data class ObjPreviewSettingsState(
         const val DEFAULT_LINE_WIDTH = 1f
         const val DEFAULT_POINT_SIZE = 3f
         const val DEFAULT_CROP_TEXTURES = false
+        const val DEFAULT_MIPMAPPING = false
         const val DEFAULT_DISPLACEMENT_QUALITY = 7.5f
     }