+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 - Automatically preview material at caret
 - Settings for MTL file editor layout separate from OBJ file editor layout
+- Setting for generating mipmaps for textures (on/off)
 
 ### Changed